% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimensionType.R
\name{check_dimensionType}
\alias{check_dimensionType}
\title{Check Dimension Type}
\usage{
check_dimensionType(dimensionType)
}
\arguments{
\item{dimensionType}{character; A vector of dimension type. Must only
contain the following values: "area", "depth", "width",
"discharge"}
}
\value{
Returns TRUE if the vector of dimension types only contain valid
values.
}
\description{
Check to determine the if a valid dimension type has been
supplied.
}
\seealso{
Other {check functions}: 
\code{\link{check_regions}()}
}
\concept{{check functions}}
