% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_dimension_graph.R
\name{region_dimension_graph}
\alias{region_dimension_graph}
\title{Region Dimension Graph}
\usage{
region_dimension_graph(regions, dimension_type)
}
\arguments{
\item{regions}{character; A vector of region names that will be
displayed on the graph. For a complete list of
regions, see levels(regional_curve$region_name)
RegionalCurve::regional_curve$region_name.}

\item{dimension_type}{character; Dimension type: "area", "depth", "width",
"discharge"}
}
\value{
A ggplot object
}
\description{
Creates a Regional Hydraulic Dimension graph for the regions
provided for the specified dimension.
}
\examples{
regions = c("Minnesota Eastern", "Minnesota Western")
dimension_type = "width"
region_dimension_graph(regions, dimension_type)

}
