% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_curves.R
\name{regional_curves}
\alias{regional_curves}
\title{Get Regional Curves}
\usage{
regional_curves()
}
\value{
A data frame of the \code{regional_curve} package data.
}
\description{
Returns a data frame of the \code{regional_curve} package data.
}
\examples{
rc <- regional_curves()

}
