% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalCurve-package.R
\docType{package}
\name{RegionalCurve-package}
\alias{RegionalCurve}
\alias{RegionalCurve-package}
\title{RegionalCurve: A Collection of Hydraulic Regional Curve Relationships}
\description{
This package contains data from a broad range of studies that delineate regional hydraulic geometry relationships. These data have been compiled into a single table to provide a comprehensive one-stop source. Accessor functions have also been included to streamline application of dimensions in other analyses.
}
\keyword{internal}
