% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHG.R
\name{RHG}
\alias{RHG}
\title{Computes regional hydraulic geometry dimensions}
\usage{
RHG(
  region,
  drainageArea,
  dimensionType = c("area", "depth", "width", "discharge")
)
}
\arguments{
\item{region}{character; The region that a dimension will be
calculated for. For a complete list of regions, see
levels(regional_curve$region_name)}

\item{drainageArea}{numeric; The upstream drainage area of the location of
the dimension (in square miles).}

\item{dimensionType}{character; Dimension type: "area", "depth", "width",
"discharge"}
}
\value{
The numeric value of the requested hydraulic dimension. The units
        of the dimensions are as follows: area: square feet, depth: feet,
        width: feet, discharge: cubic feet per second
}
\description{
Computes regional hydraulic geometry (RHG) dimension (cross
  sectional area, width, depth, discharge) from a built-in table of regional
  hydraulic geometry equation coefficients.
}
\details{
This function uses values in the \code{regional_curve} data frame
    to calculate hydraulic geometry dimensions. Monomial relationships of
    the form \eqn{y=ax^m}, known as power functions, appear as straight
    lines in a log–log graph, with the exponent (\code{m}) and constant
    (\eqn{a}) term corresponding to slope and intercept of the line
    (\eqn{y = mx + b}).
}
\examples{
# Calculate the discharge for a 200 sq mi watershed in Massachusetts.
RHG(region = "MA", drainageArea = 200, dimensionType = "discharge")

}
