% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHG_graph.R
\name{RHG_graph}
\alias{RHG_graph}
\title{Compute Hydraulic Geometry From Graph}
\usage{
RHG_graph(region, drainageArea, dimensionType)
}
\arguments{
\item{region}{character; The region that a dimension will be
calculated for. See the regional_curves$region field
for a complete list.}

\item{drainageArea}{numeric; The upstream drainage area of the location of
the dimension (in square miles).}

\item{dimensionType}{character; The dimension type: "area", "depth", "width"}
}
\value{
The value of the requested hydraulic dimension
}
\description{
Computes hydraulic dimension (cross sectional area, width,
    or depth) from a built-in table of regional hydraulic curve graph
    coordinates.
}
\examples{
# Calculate the cross sectional area for a watershed with drainage area = 1
# square mile in the Eastern United States region.
RHG_graph(region = "Eastern United States",
          drainageArea = 1,
          dimensionType = "area")

}
