% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHG_graph_coefficients.R
\name{RHG_graph_coefficients}
\alias{RHG_graph_coefficients}
\title{Compute power function coefficients}
\usage{
RHG_graph_coefficients(region, drainageArea, dimensionType)
}
\arguments{
\item{region}{character; The region that a dimension will be
calculated for. See the regional_curves$region field
for a complete list.}

\item{drainageArea}{numeric; The upstream drainage area of the location of
the dimension (in square miles).}

\item{dimensionType}{character; The dimension type: "area", "depth", "width"}
}
\value{
The value of the requested hydraulic dimension
}
\description{
Computes the regional curve power function coefficients from a table of
regional hydraulic curve coordinates contained in a built-in table.
}
\examples{

}
