% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_regions.R
\name{check_regions}
\alias{check_regions}
\title{Check Regions}
\usage{
check_regions(regions)
}
\arguments{
\item{regions}{character vector; A vector of regions listed in the
package data frame regional_curve$region_name.}
}
\value{
Returns TRUE if the \code{regions} parameter listed in the package data
frame regional_curve$region_name.
}
\description{
Checks if the parameter is listed in the package data frame
regional_curve$region_name.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\seealso{
Other {check functions}: 
\code{\link{check_dimensionType}()}
}
\concept{{check functions}}
