% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalCurve-package.R
\docType{package}
\name{RegionalCurve-package}
\alias{RegionalCurve}
\alias{RegionalCurve-package}
\title{RegionalCurve: A Collection of Hydraulic Regional Curve Relationships}
\description{
Data from studies that delineate regional hydraulic geometry relationships. These data have been compiled from a wide range of studies into a single table to provide a comprehensive one-stop source. Accessor functions have also been included to streamline the application of dimensions in other analyses and visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FluvialGeomorph/RegionalCurve}
  \item \url{https://fluvialgeomorph.github.io/RegionalCurve/}
  \item Report bugs at \url{https://github.com/FluvialGeomorph/RegionalCurve/issues}
}

}
\author{
\strong{Maintainer}: Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil} (\href{https://orcid.org/0000-0002-1465-5927}{ORCID})

Authors:
\itemize{
  \item Christopher Haring \email{Christopher.P.Haring@usace.army.mil}
}

}
\keyword{internal}
