% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHG.R
\name{RHG}
\alias{RHG}
\title{Computes hydraulic geometry dimensions}
\usage{
RHG(region, drainageArea, dimensionType)
}
\arguments{
\item{region}{character; The region that a dimension will be
calculated for. For a complete list of regions, see
levels(regional_curve$region_name)}

\item{drainageArea}{numeric; The upstream drainage area of the location of
the dimension (in square miles).}

\item{dimensionType}{character; Dimension type: "area", "depth", "width",
"discharge"}
}
\value{
The numeric value of the requested hydraulic dimension.
}
\description{
Computes hydraulic geometry dimensions
}
\examples{
# Calculate the discharge for a 200 sq mi watershed in Massachusetts.
RHG(region = "MA", drainageArea = 200, dimensionType = "discharge")

}
